pyproject.toml
[project]
name = "mushitroom-py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gpiozero>=2.0.1",
    "luma-core>=2.5.2",
    "luma-lcd>=2.11.0",
    "pillow==12.0.0",
]
